move example data to examples/example-data.tsx
