--- examples/extend-express-app/seed-data/data.ts
@@ -1,72 +1,0 @@
-export const persons = [
-  { name: 'Lucy Wroblewski' },
-  { name: 'Ches Adebayor' },
-  { name: 'Tiff Jayden' },
-  { name: 'Henrique Urrea' },
-];
-
-export const tasks = [
-  {
-    label: 'Install Keystone in local dev 🧪',
-    isComplete: true,
-    finishBy: '2021-01-01T02:30:00.000Z',
-    assignedTo: 'Lucy Wroblewski',
-    priority: 'high',
-  },
-  {
-    label: 'Model the content💡',
-    isComplete: true,
-    finishBy: '2021-01-22T05:43:51.000Z',
-    assignedTo: 'Ches Adebayor',
-    priority: 'high',
-  },
-  {
-    label: 'Architect the data schema 🔗',
-    isComplete: true,
-    finishBy: '2021-02-02T20:02:37.000Z',
-    assignedTo: 'Lucy Wroblewski',
-    priority: 'high',
-  },
-  {
-    label: 'Design the UI 💅🏼',
-    isComplete: true,
-    finishBy: '2021-02-24T22:17:07.000Z',
-    assignedTo: 'Tiff Jayden',
-    priority: 'medium',
-  },
-  {
-    label: 'Publish the content 📝',
-    isComplete: true,
-    finishBy: '2021-03-01T05:41:37.000Z',
-    assignedTo: 'Ches Adebayor',
-    priority: 'low',
-  },
-  {
-    label: 'Query content over GraphQL🔎',
-    isComplete: false,
-    finishBy: '2021-03-21T05:41:37.000Z',
-    assignedTo: 'Lucy Wroblewski',
-    priority: 'medium',
-  },
-  {
-    label: 'Implement the UI design in code 🖼',
-    isComplete: false,
-    finishBy: '2021-03-23T05:41:37.000Z',
-    assignedTo: 'Henrique Urrea',
-    priority: 'medium',
-  },
-  {
-    label: 'Deploy Keystone backend to the web ☁️',
-    isComplete: false,
-    finishBy: '2021-03-30T05:41:37.000Z',
-    assignedTo: 'Lucy Wroblewski',
-    priority: 'low',
-  },
-  {
-    label: 'Launch project 🚀',
-    isComplete: false,
-    finishBy: '2021-04-01T05:41:37.000Z',
-    assignedTo: 'Lucy Wroblewski',
-    priority: 'low',
-  },
-];
