+++ examples/singleton/seed-data.tsx
@@ -1,5 +1,7 @@
-import { authors, posts } from './data';
-import { Context } from '.keystone/types';
+import { getContext } from '@keystone-6/core/context';
+import { authors, posts } from '../example-data';
+import config from './keystone';
+import * as PrismaModule from '.prisma/client';
 
 type AuthorProps = {
   name: string;
@@ -14,17 +16,18 @@ type PostProps = {
   content: string;
 };
 
-export async function insertSeedData(context: Context) {
-  console.log(`🌱 Inserting seed data`);
+async function main() {
+  const context = getContext(config, PrismaModule);
 
+  console.log(`🌱 Inserting seed data`);
   const createAuthor = async (authorData: AuthorProps) => {
-    let author = await context.query.Author.findOne({
+    const author = await context.query.Author.findOne({
       where: { email: authorData.email },
       query: 'id',
     });
 
     if (!author) {
-      author = await context.query.Author.createOne({
+      await context.query.Author.createOne({
         data: authorData,
         query: 'id',
       });
@@ -32,7 +35,7 @@ export async function insertSeedData(context: Context) {
   };
 
   const createPost = async (postData: PostProps) => {
-    let authors = await context.query.Author.findMany({
+    const authors = await context.query.Author.findMany({
       where: { name: { equals: postData.author } },
       query: 'id',
     });
@@ -47,6 +50,7 @@ export async function insertSeedData(context: Context) {
     console.log(`👩 Adding author: ${author.name}`);
     await createAuthor(author);
   }
+
   for (const post of posts) {
     console.log(`📝 Adding post: ${post.title}`);
     await createPost(post);
@@ -54,5 +58,6 @@ export async function insertSeedData(context: Context) {
 
   console.log(`✅ Seed data inserted`);
   console.log(`👋 Please start the process with \`yarn dev\` or \`npm run dev\``);
-  process.exit();
 }
+
+main();
