+++ examples/task-manager/package.json
@@ -7,12 +7,13 @@
     "dev": "keystone dev",
     "start": "keystone start",
     "build": "keystone build",
-    "seed-data": "keystone --seed-data"
+    "seed-data": "tsx seed-data.ts"
   },
   "dependencies": {
     "@keystone-6/core": "^4.0.0"
   },
   "devDependencies": {
+    "tsx": "^3.9.0",
     "typescript": "~4.9.4"
   },
   "repository": "https://github.com/keystonejs/keystone/tree/main/examples/task-manager"
